requirements.txt
Django
openai-whisper
moviepy
Pillow
numpy
